Remove the edit doctor code to new component
